[DLMED] update according to comments
