+++ .github/workflows/pythonapp.yml
@@ -63,7 +63,7 @@ jobs:
         cat "requirements-dev.txt"
         python -m pip install -r requirements-dev.txt
         python -m pip list
-        python setup_cpp.py install
+        python setup.py install
     - name: Run quick tests (CPU ${{ runner.os }})
       run: |
         python -c 'import torch; print(torch.__version__); print(torch.rand(5,3))'
@@ -85,8 +85,7 @@ jobs:
         python -m pip uninstall -y torch torchvision
         python -m pip install torch==1.4
         python -m pip install -r requirements-dev.txt
-        python setup_cpp.py install
-        python setup_cuda.py install
+        python setup.py install
     - name: Run quick tests (GPU)
       run: |
         python -m pip list
