+++ .github/workflows/setupapp.yml
@@ -22,6 +22,7 @@ jobs:
         python -m pip install torch==1.4
         python -m pip install -r requirements-dev.txt
         python -m pip list
+        python setup.py install
     - name: Run unit tests report coverage
       run: |
         nvidia-smi
@@ -55,6 +56,7 @@ jobs:
         python -m pip install --upgrade pip wheel
         python -m pip install torch==1.4
         python -m pip install -r requirements-dev.txt
+        python setup.py install
     - name: Run quick tests CPU ubuntu
       run: |
         python -c 'import torch; print(torch.__version__); print(torch.rand(5,3))'
@@ -95,6 +97,7 @@ jobs:
         cat "requirements-dev.txt"
         python -m pip install -r requirements-dev.txt
         python -m pip list
+        python setup.py install
     - name: Run quick tests (CPU ${{ runner.os }})
       run: |
         python -c 'import torch; print(torch.__version__); print(torch.rand(5,3))'
