+++ monai/networks/extensions/lltm/lltm.cpp
@@ -98,6 +98,6 @@ std::vector<torch::Tensor> lltm_backward(
 }
 
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
-  m.def("forward", &lltm_forward, "LLTM forward");
-  m.def("backward", &lltm_backward, "LLTM backward");
+  m.def("lltm_forward", &lltm_forward, "LLTM forward");
+  m.def("lltm_backward", &lltm_backward, "LLTM backward");
 }
