+++ monai/networks/layers/simplelayers.py
@@ -10,15 +10,19 @@
 # limitations under the License.
 
 from typing import Sequence, Union
-
+import math
 import torch
-import torch.nn as nn
+from torch import nn
+from torch.autograd import Function
 import torch.nn.functional as F
-
 from monai.networks.layers.convutils import gaussian_1d, same_padding
 from monai.utils import ensure_tuple_rep
+from monai.utils import optional_import
 
-__all__ = ["SkipConnection", "Flatten", "GaussianFilter"]
+_C, _ = optional_import("monai._C")
+_C_CUDA, _ = optional_import("monai._C_CUDA")
+
+__all__ = ["SkipConnection", "Flatten", "GaussianFilter", "LLTM"]
 
 
 class SkipConnection(nn.Module):
@@ -113,3 +117,58 @@ class GaussianFilter(nn.Module):
             return self.conv_n(input=_conv(input_, d - 1), weight=kernel, padding=padding, groups=chns)
 
         return _conv(x, sp_dim - 1)
+
+
+class LLTMFunction(Function):
+    @staticmethod
+    def forward(ctx, input, weights, bias, old_h, old_cell):
+        ext = _C_CUDA if weights.is_cuda else _C
+        outputs = ext.lltm_forward(input, weights, bias, old_h, old_cell)
+        new_h, new_cell = outputs[:2]
+        variables = outputs[1:] + [weights]
+        ctx.save_for_backward(*variables)
+
+        return new_h, new_cell
+
+    @staticmethod
+    def backward(ctx, grad_h, grad_cell):
+        if grad_h.is_cuda:
+            outputs = _C_CUDA.lltm_backward(grad_h.contiguous(), grad_cell.contiguous(), *ctx.saved_tensors)
+            d_old_h, d_input, d_weights, d_bias, d_old_cell, d_gates = outputs
+        else:
+            outputs = _C.lltm_backward(grad_h.contiguous(), grad_cell.contiguous(), *ctx.saved_tensors)
+            d_old_h, d_input, d_weights, d_bias, d_old_cell = outputs
+
+        return d_input, d_weights, d_bias, d_old_h, d_old_cell
+
+
+class LLTM(nn.Module):
+    """
+    This recurrent unit is similar to an LSTM, but differs in that it lacks a forget
+    gate and uses an Exponential Linear Unit (ELU) as its internal activation function.
+    Because this unit never forgets, call it LLTM, or Long-Long-Term-Memory unit.
+    It has both C++ and CUDA implementation, automatically switch according to the
+    target device where put this module to.
+
+    Args:
+        input_features: size of input feature data
+        state_size: size of the state of recurrent unit
+
+    Referring to: https://pytorch.org/tutorials/advanced/cpp_extension.html
+    """
+
+    def __init__(self, input_features: int, state_size: int):
+        super(LLTM, self).__init__()
+        self.input_features = input_features
+        self.state_size = state_size
+        self.weights = nn.Parameter(torch.empty(3 * state_size, input_features + state_size))
+        self.bias = nn.Parameter(torch.empty(1, 3 * state_size))
+        self.reset_parameters()
+
+    def reset_parameters(self):
+        stdv = 1.0 / math.sqrt(self.state_size)
+        for weight in self.parameters():
+            weight.data.uniform_(-stdv, +stdv)
+
+    def forward(self, input, state):
+        return LLTMFunction.apply(input, self.weights, self.bias, *state)
