+++ setup.py
@@ -10,14 +10,36 @@
 # limitations under the License.
 
 from setuptools import find_packages, setup
-
+import torch
+from torch.utils.cpp_extension import BuildExtension, CppExtension, CUDAExtension
 import versioneer
 
 if __name__ == "__main__":
+    cmds = versioneer.get_cmdclass()
+    cmds.update({"build_ext": BuildExtension})
+    ext_modules = [
+        CppExtension(
+            "monai._C",
+            [
+                "monai/networks/extensions/lltm/lltm.cpp"
+            ]
+        )
+    ]
+    if torch.cuda.is_available():
+        ext_modules.append(
+            CUDAExtension(
+                "monai._C_CUDA",
+                [
+                    "monai/networks/extensions/lltm/lltm_cuda.cpp",
+                    "monai/networks/extensions/lltm/lltm_cuda_kernel.cu"
+                ],
+            )
+        )
     setup(
         version=versioneer.get_version(),
-        cmdclass=versioneer.get_cmdclass(),
+        cmdclass=cmds,
         packages=find_packages(exclude=("docs", "examples", "tests", "research")),
         zip_safe=False,
         package_data={"monai": ["py.typed"]},
+        ext_modules=ext_modules,
     )
